+++ controllers/api.js
@@ -7,8 +7,10 @@ exports.getAllTasks = function (req, res) {
 };
 
 exports.getTask = function (req, res) {
-	var task = database.filter(task => task.id == req.params.id);
-	if (task.length == 1) {
+	var task = database.filter(function (task) {
+        return task.id == req.params.id;
+    });
+	if (task.length) {
 		res.json(task[0]);
 	} else {
 		res.status(404).json({ message: 'No task with this id'});
@@ -28,8 +30,10 @@ exports.addTask = function (req, res) {
 };
 
 exports.changeTask = function (req, res) {
-	var task = database.filter(task => task.id == req.params.id);
-	if (task.length == 1) {
+	var task = database.filter(function (task) {
+        return task.id == req.params.id;
+    });
+	if (task.length) {
 		task[0].text = req.body.text;
 		res.status(200).json({ message: 'OK' });
 	} else {
@@ -38,9 +42,10 @@ exports.changeTask = function (req, res) {
 };
 
 exports.deleteTask = function (req, res) {
-	for (var i = 0;
-		i < database.length && database[i].id != req.params.id;
-		i++) {}
+	var i = 0;
+    while (i < database.length && database[i].id != req.params.id) {
+        i++;
+    }
 	database.splice(i, 1);
 	res.status(200).json({ message: 'OK' });
 };
