+++ public/scripts/index.jsx
@@ -24,11 +24,17 @@ function createTaskOnFront(task) {
 }
 
 function showLoading() {
-    document.getElementsByClassName('loading-image')[0].style.display = 'inline';
+	var loadingImages = document.getElementsByClassName('loading-image');
+	if (loadingImages.length) {
+		loadingImages[0].style.display = 'inline';
+	}
 }
 
 function hideLoading() {
-    document.getElementsByClassName('loading-image')[0].style.display = 'none';
+    var loadingImages = document.getElementsByClassName('loading-image');
+	if (loadingImages.length) {
+		loadingImages[0].style.display = 'none';
+	}
 }
 
 function getNewTasks() {
@@ -46,143 +52,6 @@ function getTasks() {
         tasks.forEach(createTaskOnFront);
     });
 }
-
-var TaskItem = React.createClass({
-    getInitialState: function () {
-        var taphandler = new TapHandler();
-        taphandler.swipeLeftCB = this.onSwipeLeft;
-        taphandler.swipeRightCB = this.onSwipeRight;
-        return {
-            taphandler: taphandler,
-            edit: false,
-            isDelete: false,
-            id: this.props.id,
-            text: this.props.text
-        };
-    },
-    
-    changeText: function (event) {
-        this.setState({text: event.target.value});
-    },
-    
-    updateTask: function (event) {
-        changeTask(this.state.id, this.state.text);
-    },
-    
-    onSwipeLeft: function () {
-        this.setState({ isDelete: true });
-    },
-    
-    onSwipeRight: function () {
-        this.setState({ isDelete: false });
-    },
-    
-    onClick: function () {
-        this.setState({ edit: !this.state.edit });
-    },
-    
-    onTextAreaClick: function (event) {
-        event.stopPropagation();
-    },
-    
-    onDeleteTask: function (event) {
-        event.stopPropagation();
-        deleteTask(this.state.id, (function () {
-            this.props.deleteCB();
-        }).bind(this));
-    },
-    
-   render: function () {
-        var innerContent;
-        if (this.state.edit) {
-            innerContent =  (
-                <div className="task-item__editor">
-                    <textarea 
-                    className="task-item__editor__area" 
-                    value={this.state.text}
-                    onChange={this.changeText}
-                    onClick={this.onTextAreaClick}>
-                    </textarea>
-                    <br />
-                    <button 
-                    onClick={this.updateTask}
-                    className="task-item__save-button">
-                        Сохранить
-                    </button>
-                </div>
-            );
-        } else {
-            innerContent = <div className="task-item__name">{this.state.text}</div>;
-        }
-        var className = "task-item task-item_edit_false task-item_delete_" +
-            this.state.isDelete.toString();
-        return (
-            <div className={className}
-                onTouchStart={this.state.taphandler.touchStart.bind(this.state.taphandler)}
-                onTouchMove={this.state.taphandler.touchMove.bind(this.state.taphandler)}
-                onTouchEnd={this.state.taphandler.touchEnd.bind(this.state.taphandler)}
-                onClick={this.onClick}
-                >
-                {innerContent}
-                <div className="task-item__delete-task">
-                <img 
-                src="/images/delete.png"
-                width="30" 
-                height="30"
-                onClick={this.onDeleteTask}
-                />
-            </div>
-            </div>
-        );
-    }
-    
-});
-
-var TaskList = React.createClass({
-
-    getInitialState: function () {
-        return {tasks: []};
-    },
-    
-    componentDidMount: function () {
-        addTaskCB = this.addTask;
-        clearTasksCB = this.clearTasks;
-        setTimeout(completelyLoaded(), 0);
-    },
-    
-    addTask: function (task) {
-        this.setState({ tasks: this.state.tasks.concat(task)});
-    },
-    
-    clearTasks: function () {
-        this.setState({ tasks: [] });
-    },
-    
-    getDeleteCB: function (id) {
-        return (function () {
-            var newTasks = this.state.tasks.filter(function (task) {
-                return task.id != id
-            });
-            this.setState({ tasks: newTasks });
-        }).bind(this);
-    },
-
-    render: function () {
-        var items = [];
-        this.state.tasks.forEach((function (task) {
-            items.push(<TaskItem
-                key={task.id}
-                id={task.id}
-                text={task.text}
-                deleteCB={this.getDeleteCB(task.id)}
-                />);
-        }).bind(this));
-        return <div>{items}</div>;
-    }
-});
-
-var addTaskCB;
-var clearTasksCB;
 
 function completelyLoaded() {
     document.getElementsByClassName('task-item__add-button')[0]
